refactor: split pkg/directory/interfaces/schema.go into 5 focused files (Issue #453)
The 1,367-line schema.go mixed four distinct responsibilities: type
definitions, mapper implementation, transformer implementations, and
validator implementations. This mixed-concern structure made it hard
to navigate and understand what each layer does at a glance.

Extracted into five focused files with clear single responsibilities:
- schema_types.go       (~232 lines) — All type defs, constants, interfaces
- schema.go             (~601 lines) — Mapper construction, field mapping,
                                       object conversion (User/Group)
- schema_validation.go  (~298 lines) — Object/field validation and conflict
                                       resolution
- schema_transformers.go (~196 lines) — 7 DataTransformer implementations
- schema_validators.go  (~100 lines) — 4 FieldValidator implementations

The transformers and validators are self-contained implementations of
interfaces defined in schema_types.go. Splitting them clarifies that
they are leaf implementations with no dependencies on the mapper.
Validation and conflict resolution logic is similarly cohesive and
benefits from being grouped separately from field mapping operations.

No public API changes — all types, interfaces, and functions remain
identical. Only file boundaries changed. All tests pass unchanged.

Changes:
- Extract type definitions and interface declarations to schema_types.go
- Extract validation/conflict resolution methods to schema_validation.go
- Extract DataTransformer implementations to schema_transformers.go
- Extract FieldValidator implementations to schema_validators.go
- Update schema.go file header to document the 5-file organization
- Remove strconv from schema.go imports (only used in ParseBooleanTransformer)

Fixes #453
